+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.51
 - [common] Override MinecraftVersion#toString
+- [common] Add MinecraftVersion#getByProtocolVersion
 - [common] Add 26.1.x to MinecraftVersion
 
 ## 0.0.50